Performed requested changes found in PR
