+++ api/app/services/threads/embedding_thread.py
@@ -1,30 +1,24 @@
 import gc
 import logging
 import threading
-from datetime import datetime
+from datetime import datetime, timezone
 from functools import partial
-from time import sleep
-from typing import Callable
+from typing import Callable, Literal
 
 import numpy as np
 import torch
 from app.config import settings
-from app.helper import (
-    _perform_request,
-    parse_asset_date,
-    translate_datetime_to_aiod_params,
-)
 from app.models.asset_collection import AssetCollection
 from app.schemas.enums import AssetType
 from app.schemas.request_params import RequestParams
+from app.services.aiod import recursive_aiod_asset_fetch
 from app.services.database import Database
-from app.services.embedding_store import EmbeddingStore, Milvus_EmbeddingStore
+from app.services.embedding_store import EmbeddingStore, MilvusEmbeddingStore
 from app.services.inference.model import AiModel
 from app.services.inference.text_operations import (
     ConvertJsonToString,
-    HuggingFaceDatasetExtractMedatada,
+    HuggingFaceDatasetExtractMetadata,
 )
-from requests.exceptions import HTTPError
 from torch.utils.data import DataLoader
 
 job_lock = threading.Lock()
@@ -54,7 +48,7 @@ async def compute_embeddings_for_aiod_assets_wrapper(
 async def compute_embeddings_for_aiod_assets(first_invocation: bool) -> None:
     model = AiModel(device=AiModel.get_device())
     database = Database()
-    embedding_store = await Milvus_EmbeddingStore.init()
+    embedding_store = await MilvusEmbeddingStore.init()
 
     try:
         asset_types = settings.AIOD.ASSET_TYPES
@@ -83,7 +77,7 @@ async def compute_embeddings_for_aiod_assets(first_invocation: bool) -> None:
             meta_extract_types = settings.AIOD.ASSET_TYPES_FOR_METADATA_EXTRACTION
             if settings.MILVUS.EXTRACT_METADATA and asset_type in meta_extract_types:
                 extract_metadata_func = partial(
-                    HuggingFaceDatasetExtractMedatada.extract_huggingface_dataset_metadata,
+                    HuggingFaceDatasetExtractMetadata.extract_huggingface_dataset_metadata,
                     asset_type=asset_type,
                 )
 
@@ -92,7 +86,7 @@ async def compute_embeddings_for_aiod_assets(first_invocation: bool) -> None:
                 stringify_function=partial(
                     ConvertJsonToString.extract_relevant_info, asset_type=asset_type
                 ),
-                extract_medatadata_function=extract_metadata_func,
+                extract_metadata_function=extract_metadata_func,
                 embedding_store=embedding_store,
                 database=database,
                 asset_collection=asset_collection,
@@ -108,13 +102,12 @@ async def compute_embeddings_for_aiod_assets(first_invocation: bool) -> None:
 def process_aiod_assets_wrapper(
     model: AiModel,
     stringify_function: Callable[[dict], str],
-    extract_medatadata_function: Callable[[dict], dict] | None,
+    extract_metadata_function: Callable[[dict], dict] | None,
     embedding_store: EmbeddingStore,
     database: Database,
     asset_collection: AssetCollection,
     asset_type: AssetType,
 ) -> None:
-    asset_url = settings.AIOD.get_assets_url(asset_type)
     existing_doc_ids_from_past = embedding_store.get_all_document_ids(asset_type)
     newly_added_doc_ids = []
 
@@ -148,7 +141,7 @@ def process_aiod_assets_wrapper(
 
     while True:
         assets_to_add, asset_ids_to_remove = get_assets_to_add_and_delete(
-            asset_url,
+            asset_type,
             url_params,
             existing_doc_ids_from_past=existing_doc_ids_from_past,
             newly_added_doc_ids=newly_added_doc_ids,
@@ -175,8 +168,8 @@ def process_aiod_assets_wrapper(
             asset_ids = [str(obj["identifier"]) for obj in assets_to_add]
 
             metadata = [{} for _ in assets_to_add]
-            if extract_medatadata_function is not None:
-                metadata = [extract_medatadata_function(obj) for obj in assets_to_add]
+            if extract_metadata_function is not None:
+                metadata = [extract_metadata_function(obj) for obj in assets_to_add]
 
             data = [
                 (obj, id, meta)
@@ -211,14 +204,14 @@ def process_aiod_assets_wrapper(
 
 
 def get_assets_to_add_and_delete(
-    url: str,
+    asset_type: AssetType,
     url_params: RequestParams,
     existing_doc_ids_from_past: list[str],
     newly_added_doc_ids: list[str],
     last_db_sync_datetime: datetime | None,
 ) -> tuple[list[dict] | None, list[str] | None]:
     mark_recursions = []
-    assets = recursive_fetch(url, url_params, mark_recursions)
+    assets = recursive_aiod_asset_fetch(asset_type, url_params, mark_recursions)
 
     if len(assets) == 0 and len(mark_recursions) == 0:
         # We have reached the end of the AIoD database
@@ -232,7 +225,7 @@ def get_assets_to_add_and_delete(
 
     asset_ids = [str(obj["identifier"]) for obj in assets]
     modified_dates = np.array(
-        [parse_asset_date(obj, none_value="now") for obj in assets]
+        [parse_aiod_asset_date(obj, none_value="now") for obj in assets]
     )
 
     # new assets to store that we have never encountered before
@@ -260,44 +253,18 @@ def get_assets_to_add_and_delete(
     return assets_to_add, asset_ids_to_del
 
 
-def recursive_fetch(
-    url: str, url_params: RequestParams, mark_recursions: list[int]
-) -> list:
-    try:
-        sleep(settings.AIOD.JOB_WAIT_INBETWEEN_REQUESTS_SEC)
-        queries = _build_real_url_queries(url_params)
-        response = _perform_request(url, queries)
-        data = response.json()
-    except HTTPError as e:
-        if e.response.status_code != 500:
-            raise e
-        if url_params.limit == 1:
-            return []
+def parse_aiod_asset_date(
+    asset: dict,
+    field: str = "date_modified",
+    none_value: Literal["none", "now", "zero"] = "none",
+) -> datetime | None:
+    string_time = asset.get("aiod_entry", {}).get(field, None)
+    if string_time is None:
+        if none_value == "none":
+            return None
+        if none_value == "now":
+            return datetime.now(tz=timezone.utc)
+        if none_value == "zero":
+            return datetime.fromtimestamp(0, tz=timezone)
 
-        mark_recursions.append(1)
-        first_half_limit = url_params.limit // 2
-        second_half_limit = url_params.limit - first_half_limit
-
-        first_half = recursive_fetch(
-            url, url_params.new_page(limit=first_half_limit), mark_recursions
-        )
-        second_half = recursive_fetch(
-            url,
-            url_params.new_page(
-                offset=url_params.offset + first_half_limit, limit=second_half_limit
-            ),
-            mark_recursions,
-        )
-        return first_half + second_half
-
-    return data
-
-
-def _build_real_url_queries(url_params: RequestParams) -> dict:
-    return {
-        "schema": "aiod",
-        "offset": url_params.offset,
-        "limit": url_params.limit,
-        "date_modified_after": translate_datetime_to_aiod_params(url_params.from_time),
-        "date_modified_before": translate_datetime_to_aiod_params(url_params.to_time),
-    }
+    return datetime.fromisoformat(string_time).replace(tzinfo=timezone.utc)
